Reorganized Test folder - moved utilities script to `Tests/Utilities` folder, assembly, and namespace - restructured `Tests/Resources` folder to contain and organize content appropriately - remove now-unused testing plugins - adjusted assembly and script references
