--- Tests/Plugins.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c0ee5b853950f0940993fb357b65a6e3
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
