--- Tests/Plugins/NSubstitute/README.md
@@ -1,3 +1,0 @@
-#NSubstitute v4.2.1
-
-[Website](https://www.nuget.org/packages/NSubstitute/4.2.1)
