--- Tests/Plugins/NSubstitute/README.md.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: fbb7b9b4e8440c64a92eaf7c88e6ee96
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
