+++ Tests/Resources/Bci4Kids.BciEssentials.Tests.Resources.asmdef.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: faf67fcf8b51f9e479d0fdea21626618
+guid: 75aa2154b3fedd141910912cdd73270e
 AssemblyDefinitionImporter:
   externalObjects: {}
   userData:
