--- Tests/Resources/Editor/BCIControllerBehaviorExtensions.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 9980febd96444744ab52c9e277dd405f
-timeCreated: 1675017266
