--- Tests/Resources/Editor/BCIControllerExtensions.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 31996ab52d3647b9b5ef7325ddb52d34
-timeCreated: 1675018284
