--- Tests/Resources/Editor/LoggingHelpers.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 1b41078e5abe4642942725251a47874f
-timeCreated: 1672115192
