--- Tests/Resources/Editor/ReflectionHelpers.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 5c2130adac6641a98eb16100139cb89f
-timeCreated: 1675017531
