--- Tests/Resources/Editor/ResourceLibrary.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 2a929b4697a641428c26ebe3197179c5
-timeCreated: 1672170019
