+++ Tests/Resources/ResourceLibraryAsset.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 437f542230e2e8f489eee74745704312
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
