--- Tests/Resources/Runtime.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3da41515930f9e949a3217707fd157a0
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
