--- Tests/Resources/Runtime/Scripts.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: f65ed092a93c44efbe29f918a7007453
-timeCreated: 1672169872
