--- Tests/Resources/Runtime/Scripts/CoroutineRunner.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 3665c038736942bf88356b4b2d5c1091
-timeCreated: 1672622366
