+++ Tests/Resources/Scenes/EmptyScene.unity
