+++ Tests/Resources/Scenes/EmptyScene.unity.meta
