--- Tests/Runtime/EditorSceneLoader.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: bcf86abd9a0f4890834c26f652945790
-timeCreated: 1672124720
