+++ Tests/Utilities/Bci4Kids.BciEssentials.Tests.Utlilities.asmdef.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: a0c9ed9fb854f5540bdd27013fa68b66
-folderAsset: yes
-DefaultImporter:
+guid: 11d1a408a591325449fb595bf59a735b
+AssemblyDefinitionImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
