+++ Tests/Utilities/CoroutineRunner.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d603d5694d02cb5418195d860e68be38
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
