+++ Tests/Utilities/LSLRawStreamReader.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 268910c83fdd99f409086622c7580542
+guid: ca67c97418d68cd4086477a223e9439f
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
